ecommerce/requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.1
django-crispy-forms==2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.7.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
